+++ CHANGELOG.md
@@ -5,15 +5,21 @@ The format is based on [Keep a Changelog(https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning(https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Add `--timeout` option to the WeasyPrint command-line call by default. This improves consistency with the internal process timeout already applied by Symfony Process. If you're running WeasyPrint inside a worker, queue, or other timeout-managed environment, you can disable it using `$pdf->disableTimeout()` or `$pdf->setTimeout(null)`. (#15)
+- Add `disableTimeout()` method to easily disable the new CLI timeout behavior
+
+### Security
+- Update `symfony/process` minimal version to mitigate [CVE-2024-51736](https://github.com/advisories/GHSA-qq5c-677p-737q)
 
 ## 1.5.0 - 2024-11-04
 ### Added
 - Support WeasyPrint 63.0 new `srgb` option
-- Added support for PHP 8.4
+- Add support for PHP 8.4
 
 ## 1.4.0 - 2023-11-20
 ### Changed
-- Added support for Symfony 7.0 and PHP 8.3
+- Add support for Symfony 7.0 and PHP 8.3
 
 ## 1.3.0 - 2023-10-07
 ### Added
@@ -39,7 +45,7 @@ and this project adheres to [Semantic Versioning(https://semver.org/spec/v2.0.0.
 
 ## 1.0.0 - 2023-01-16
 ### Fixed
-- Fixed handling of repeatable options (attachment and stylesheet)
+- Fix handling of repeatable options (attachment and stylesheet)
 
 ### Changed
 - Bump symfony/process up to ^6.2